requirements.txt
fastapi>=0.124.0,<0.125.0
uvicorn[standard]
openai>=1.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0
